Dynamically load blog posts on blog
Blog posts linked too another HTML file, which after clicking on the
link, brought you to another page. This causes the user to have to load
a new page. Instead, clicking on the blog links dynamically loads the
posts into the page, causing the user to not have to load a new page to
see posts.

Signed-off-by: Kien Ha <kienha9922@gmail.com>